Update device pixel ratio on screen change
This patch makes QWebPage listen to QWindow::screenChangedEvent
and update the device pixel ratio when it is received.

Task-number: QTBUG-36190

+++ Source/WebKit/qt/WebCoreSupport/QWebPageAdapter.cpp
@@ -1337,7 +1337,7 @@ QWebPageAdapter::ViewportAttributes QWebPageAdapter::viewportAttributesForSize(c
     return result;
 }
 
-void QWebPageAdapter::setDevicePixelRatio(int devicePixelRatio)
+void QWebPageAdapter::setDevicePixelRatio(float devicePixelRatio)
 {
     page->setDeviceScaleFactor(devicePixelRatio);
 }
