+++ Source/WebKit/qt/WebCoreSupport/QWebPageAdapter.h
@@ -374,7 +374,7 @@ public:
     };
 
     ViewportAttributes viewportAttributesForSize(const QSize& availableSize, const QSize& deviceSize) const;
-    void setDevicePixelRatio(int devicePixelRatio);
+    void setDevicePixelRatio(float devicePixelRatio);
 
     QWebSettings *settings;
